Tarrin and Great war super events
This git is DANOIR APPROVED
Music for great war : Symphony No. 4 in A Major, Op. 90 aka Italian symphony
Image for great  war, 2 panzer 3 in on the field with their commanders sitting on top
Music for the Tarrin war : Danse Macabre(Saint-Saëns) aka the Dance of Death
image for the Tarrin war : 2 S35 and an artillery piece with a guy walking across the frame