Tassen Buffing His Country
You can call this a "GRI Buff" but this is currently how the focus works anyway. As soon as the focus completes, you can click the two annex decisions instantly before YAL becomes free. All this does it make it not dependent on me pausing the game and clicking the decisions.